[Front] blinking of Position fix --> migration of Positions and PositionsLines to tsx (#2843)
