--- opencti-platform/opencti-front/src/private/components/locations/positions/PositionsLines.js
@@ -1,137 +1,0 @@
-import React, { Component } from 'react';
-import * as PropTypes from 'prop-types';
-import { graphql, createPaginationContainer } from 'react-relay';
-import { pathOr } from 'ramda';
-import ListLinesContent from '../../../../components/list_lines/ListLinesContent';
-import { PositionLine, PositionLineDummy } from './PositionLine';
-import { setNumberOfElements } from '../../../../utils/Number';
-
-const nbOfRowsToLoad = 50;
-
-class PositionsLines extends Component {
-  componentDidUpdate(prevProps) {
-    setNumberOfElements(
-      prevProps,
-      this.props,
-      'positions',
-      this.props.setNumberOfElements.bind(this),
-    );
-  }
-
-  render() {
-    const { initialLoading, dataColumns, relay, paginationOptions } = this.props;
-    return (
-      <ListLinesContent
-        initialLoading={initialLoading}
-        loadMore={relay.loadMore.bind(this)}
-        hasMore={relay.hasMore.bind(this)}
-        isLoading={relay.isLoading.bind(this)}
-        dataList={pathOr([], ['positions', 'edges'], this.props.data)}
-        globalCount={pathOr(
-          nbOfRowsToLoad,
-          ['positions', 'pageInfo', 'globalCount'],
-          this.props.data,
-        )}
-        LineComponent={<PositionLine />}
-        DummyLineComponent={<PositionLineDummy />}
-        dataColumns={dataColumns}
-        nbOfRowsToLoad={nbOfRowsToLoad}
-        paginationOptions={paginationOptions}
-      />
-    );
-  }
-}
-
-PositionsLines.propTypes = {
-  classes: PropTypes.object,
-  paginationOptions: PropTypes.object,
-  dataColumns: PropTypes.object.isRequired,
-  data: PropTypes.object,
-  relay: PropTypes.object,
-  initialLoading: PropTypes.bool,
-  setNumberOfElements: PropTypes.func,
-};
-
-export const positionsLinesQuery = graphql`
-  query PositionsLinesPaginationQuery(
-    $search: String
-    $count: Int!
-    $cursor: ID
-    $orderBy: PositionsOrdering
-    $orderMode: OrderingMode
-    $filters: [PositionsFiltering]
-  ) {
-    ...PositionsLines_data
-      @arguments(
-        search: $search
-        count: $count
-        cursor: $cursor
-        orderBy: $orderBy
-        orderMode: $orderMode
-        filters: $filters
-      )
-  }
-`;
-
-export default createPaginationContainer(
-  PositionsLines,
-  {
-    data: graphql`
-      fragment PositionsLines_data on Query
-      @argumentDefinitions(
-        search: { type: "String" }
-        count: { type: "Int", defaultValue: 25 }
-        cursor: { type: "ID" }
-        orderBy: { type: "PositionsOrdering", defaultValue: name }
-        orderMode: { type: "OrderingMode", defaultValue: asc }
-        filters: { type: "[PositionsFiltering]" }
-      ) {
-        positions(
-          search: $search
-          first: $count
-          after: $cursor
-          orderBy: $orderBy
-          orderMode: $orderMode
-          filters: $filters
-        ) @connection(key: "Pagination_positions") {
-          edges {
-            node {
-              id
-              name
-              description
-              ...PositionLine_node
-            }
-          }
-          pageInfo {
-            endCursor
-            hasNextPage
-            globalCount
-          }
-        }
-      }
-    `,
-  },
-  {
-    direction: 'forward',
-    getConnectionFromProps(props) {
-      return props.data && props.data.positions;
-    },
-    getFragmentVariables(prevVars, totalCount) {
-      return {
-        ...prevVars,
-        count: totalCount,
-      };
-    },
-    getVariables(props, { count, cursor }, fragmentVariables) {
-      return {
-        search: fragmentVariables.search,
-        count,
-        cursor,
-        orderBy: fragmentVariables.orderBy,
-        orderMode: fragmentVariables.orderMode,
-        filters: fragmentVariables.filters,
-      };
-    },
-    query: positionsLinesQuery,
-  },
-);
